gm_comp: fix undefined symbol while linking
This is caused by functions in gm_error.cc using 'int errno' as
function parameter, while the header defines the parameter as
'int errnumber'. This is a problem, as errno.h is included which
defines errno to be:

 #define errno (*__errno_location ())

which effectively changes the function parameter in the gm_error.cc
file to take a function pointer as the first argument and therefore
the other code doesn't find the symbols.

Signed-off-by: Reto Achermann <rachermann@vmware.com>